Split up index.ios.js into separate files
Moved Browser to components/BrowserView to componentize it for inclusion in both iOS and Android apps. Moved AppRegistry/setup code to setup.js to add redux and other middleware in future patches. Fixed ReactDefaultsPrefs bug when running app from repo (https://github.com/facebook/react-native/issues/7719)
