+++ package.json
@@ -6,7 +6,7 @@
     "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
-    "react": "^15.0.2",
+    "react": "15.0.2",
     "react-native": "^0.26.2"
   }
 }
